refactor: seperation of concerns
