+++ src/showdown/auto_replay.ts
